refactor tidal data parsing script
